fix(share): fix share button unresponsive on Overview for iOS and Android
